+++ firmware/controllers/actuators/gppwm/gppwm.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "engine.h"
+#include "engine_ptr.h"
 
 void initGpPwm(DECLARE_ENGINE_PARAMETER_SIGNATURE);
 void updateGppwm();
