+++ firmware/controllers/core/common_headers.h
@@ -27,6 +27,7 @@
 #include "efitime.h"
 
 #ifdef __cplusplus
+#include "engine_ptr.h"
 #include "datalogging.h"
 #include "loggingcentral.h"
 #include "cli_registry.h"
@@ -72,34 +73,6 @@
  */
 #define DISPLAY_SENSOR(x) {}
 #define DISPLAY_IF(x) x
-
-#if EFI_UNIT_TEST
-
-#define DECLARE_ENGINE_PTR                                 \
-	Engine *engine = nullptr;                              \
-	engine_configuration_s *engineConfiguration = nullptr; \
-	persistent_config_s *config = nullptr;
-
-
-#define INJECT_ENGINE_REFERENCE(x)                  \
-	(x)->engine = engine;                           \
-	(x)->engineConfiguration = engineConfiguration; \
-	(x)->config = config;
-
-#else // EFI_UNIT_TEST
-
-#define DECLARE_ENGINE_PTR
-
-#define INJECT_ENGINE_REFERENCE(x) {}
-
-#endif // EFI_UNIT_TEST
-
-#define EXPAND_Engine \
-	    engine_configuration_s *engineConfiguration = engine->engineConfigurationPtr; \
-		persistent_config_s *config = engine->config; \
-		(void)engineConfiguration; \
-		(void)config; \
-		
 
 #ifndef EFI_ACTIVE_CONFIGURATION_IN_FLASH
 // We store a special changeable copy of configuration is RAM, so we can just compare them
