+++ firmware/controllers/global_shared.h
@@ -29,6 +29,7 @@
  */
 
 #include "global.h"
+#include "engine_ptr.h"
 
 #define EXTERN_ENGINE_CONFIGURATION \
 		extern engine_configuration_s *engineConfiguration; \
@@ -63,25 +64,6 @@
 		extern Engine *engine; \
 		EXTERN_CONFIG \
 		extern EnginePins enginePins \
-
-// See also DECLARE_ENGINE_PTR
-// See also INJECT_ENGINE_REFERENCE
-
-
-// Use this macro to declare a function which only takes magic references
-#define DECLARE_ENGINE_PARAMETER_SIGNATURE void
-// Use this version of the macro as the suffix if method has other parameters
-#define DECLARE_ENGINE_PARAMETER_SUFFIX
-// Pass this if only magic references are needed
-#define PASS_ENGINE_PARAMETER_SIGNATURE
-// Pass this after some other parameters are passed
-#define PASS_ENGINE_PARAMETER_SUFFIX
-
-// these macro are used when we should not have visibility to 'engine'
-#define DECLARE_CONFIG_PARAMETER_SIGNATURE void
-#define DECLARE_CONFIG_PARAMETER_SUFFIX
-#define PASS_CONFIG_PARAMETER_SIGNATURE
-#define PASS_CONFIG_PARAMETER_SUFFIX
 
 #define ENGINE(x) ___engine.x
